added strategy to paddle, fixed keylistener direction, small tweaks

+++ src/main/java/pong/Environment.java
@@ -1,8 +1,8 @@
 package pong;
 
 public class Environment {
-    public static final int WIDTH = 700;
-    public static final int HEIGHT = 500;
+    public static final double WIDTH = 700;
+    public static final double HEIGHT = 500;
 
     private final Paddle user;
     private final Paddle opponent;
